Correct ID variable - tokenId should be itemId
NFTs objects should have itemId propriety and not tokenId propriety.
Creating a token but not listing it will defer the itemId and tokenId, fetching the marketplace based on tokenId will result in the wrong token being fetched.

For example, you might have created 18 tokens (tokenId.last = 18), but listed only 5. The 4th tokenId will not correspond to the 4th itemId.